requirements.txt
fastapi
uvicorn
requests
pandas
numpy
scikit-learn
lightgbm
xgboost
shap
pyyaml
python-dotenv
streamlit
plotly
sqlalchemy
celery
redis